multiple page routes added
