+++ src/App.css
@@ -3,73 +3,13 @@
   height: 100%;
 }
 
-a {
-  text-decoration: none;
-  color: white;
-  font-weight: bold;
-}
-
-.Nav {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 10px 20px;
-  background-color: #f5913e;
-}
-
-.NavItems {
-  display: flex;
-  align-items: center;
-  gap: 5px;
-  list-style: none;
-}
 
 .Main {
   padding: 20px 20px;
 }
 
-.AboutDesc {
-  margin-top: 20px;
-}
-
-.CalculatorContainer {
-  width: 30%;
-  margin: 50px auto 0;
-}
-
-.Screen {
-  background-color: #858694;
-  padding: 15px 10px 6px;
-  text-align: end;
-  font-size: 25px;
-  color: #fff;
-  font-weight: bold;
-  width: 100%;
-}
-
 .span-2-col {
   grid-column: 1 / 3;
-}
-
-.keypad-button {
-  padding: 15px 15px;
-  text-align: center;
-  width: 100%;
-  border-right: 1px solid #999;
-  border-bottom: 1px solid #999;
-}
-
-.primary-button {
-  background-color: #e0e0e0;
-}
-
-.secondary-button {
-  background-color: #f5913e;
-}
-
-.KeypadContainer {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
 }
 
 .ErrorMessage {
