+++ src/App.js
@@ -1,41 +1,14 @@
-import { useEffect, useState } from 'react';
+import React from 'react';
 import './App.css';
-import Calculator from './components/Calculator';
 import Header from './components/Header';
-import Qoute from './components/Qoute';
-import Home from './components/Home';
+import AllRoutes from './routes/index';
 
 function App() {
-  const [tab, setTab] = useState();
-  useEffect(() => {
-    setTab('Home');
-  }, []);
-
   return (
     <div className="App">
-      <Header tabChangeHandler={setTab} />
+      <Header />
       <main className="Main">
-        {
-        tab === 'Home' && (
-          <section className="HomeContainer">
-            <Home />
-          </section>
-        )
-      }
-        {
-        tab === 'Calculator' && (
-          <section className="CalculatorContainer">
-            <Calculator />
-          </section>
-        )
-      }
-        {
-        tab === 'Quote' && (
-          <section className="QuoteContainer">
-            <Qoute />
-          </section>
-        )
-      }
+        <AllRoutes />
       </main>
     </div>
   );
