+++ src/components/Header.jsx
@@ -1,20 +1,17 @@
 import React from 'react';
-
+import { Link } from 'react-router-dom';
+import Style from './Header.module.css';
 /* eslint-disable react/prop-types, jsx-a11y/no-noninteractive-element-interactions,
 jsx-a11y/click-events-have-key-events */
 
-export default function Header({ tabChangeHandler }) {
-  const navigateToUrl = ($e, tab) => {
-    $e.preventDefault();
-    return tabChangeHandler(tab);
-  };
+export default function Header() {
   return (
-    <nav className="Nav">
-      <a href="/">Calculator</a>
-      <ul className="NavItems">
-        <li><a onClick={($e) => navigateToUrl($e, 'Home')} href="/">Home</a></li>
-        <li><a onClick={($e) => navigateToUrl($e, 'Calculator')} href="/">Calculator</a></li>
-        <li><a onClick={($e) => navigateToUrl($e, 'Quote')} href="/">Quote</a></li>
+    <nav className={`${Style.Nav}`}>
+      <Link className={`${Style.Link}`} to="/">Calculator</Link>
+      <ul className={`${Style.NavItems}`}>
+        <li><Link className={`${Style.Link}`} to="/">Home</Link></li>
+        <li><Link className={`${Style.Link}`} to="/calculator">Calculator</Link></li>
+        <li><Link className={`${Style.Link}`} to="/qoute">Quote</Link></li>
       </ul>
     </nav>
   );
