+++ src/components/Header.module.css
@@ -1,0 +1,21 @@
+ .Link {
+    text-decoration: none;
+    color: white;
+    font-weight: bold;
+  }
+  
+  .Nav {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 15px 20px;
+    background-color: #f5913e;
+  }
+  
+  .NavItems {
+    display: flex;
+    align-items: center;
+    gap: 20px;
+    list-style: none;
+  }
+
