+++ src/pages/home/Home.jsx
@@ -1,16 +1,17 @@
 import React from 'react';
+import Style from './Home.module.css';
 
 export default function Home() {
   return (
-    <div>
+    <div className={`${Style.HomeContainer}`}>
       <h1>Welcome to our page.</h1>
-      <p className="AboutDesc">
+      <p className={`${Style.AboutDesc}`}>
         Lorem ipsum, dolor sit amet consectetur adipisicing elit.
         Ratione dolor, omnis, distinctio illum dolorum molestiae et,
         ipsa cumque cum culpa quam libero?
         Nihil eos aliquam rem repellat est ab doloribus.
       </p>
-      <p className="AboutDesc">
+      <p className={`${Style.AboutDesc}`}>
         Lorem ipsum dolor sit amet consectetur adipisicing elit.
         Eaque iste repudiandae ex natus ratione fugiat mollitia laboriosam est qui,
         rem similique ullam dolorem, architecto eligendi
