+++ src/pages/home/Home.module.css
@@ -1,0 +1,10 @@
+.red {
+    color: red;
+}
+.HomeContainer{
+    padding: 10px 10px;
+}
+
+.AboutDesc {
+    margin-top: 20px;
+  }
